+++ simulator/src/main.cpp
@@ -6,7 +6,7 @@ int main(int argc, char *argv[]) {
     QApplication app(argc, argv);
     QApplication::setApplicationName("BC2 Cold Wallet Simulator");
     QApplication::setOrganizationName("BC2 Cold Wallet");
-    QApplication::setApplicationVersion("0.17.4");
+    QApplication::setApplicationVersion("0.18.0");
 
     MainWindow window;
     window.show();